+++ lib/bolt/core/src/context/service.rs
@@ -862,12 +862,10 @@ impl ServiceContextData {
 			));
 		}
 
-		{
-			env.push((
-				"CRDB_MIN_CONNECTIONS".into(),
-				self.config().cockroachdb.min_connections.to_string(),
-			));
-		}
+		env.push((
+			"CRDB_MIN_CONNECTIONS".into(),
+			self.config().cockroachdb.min_connections.to_string(),
+		));
 
 		if self.depends_on_clickhouse() {
 			let clickhouse_data = terraform::output::read_clickhouse(&project_ctx).await;