+++ svc/pkg/job-run/worker/src/workers/create/mod.rs
@@ -235,7 +235,7 @@ async fn write_to_db_before_run(
 	token_session_id: Uuid,
 ) -> GlobalResult<()> {
 	sql_execute!(
-		[ctx]
+		[ctx, @tx tx]
 		"
 		INSERT INTO db_job_state.runs (run_id, region_id, create_ts, token_session_id)
 		VALUES ($1, $2, $3, $4)
@@ -248,7 +248,7 @@ async fn write_to_db_before_run(
 	.await?;
 
 	sql_execute!(
-		[ctx]
+		[ctx, @tx tx]
 		"INSERT INTO db_job_state.run_meta_nomad (run_id) VALUES ($1)",
 		run_id,
 	)
@@ -278,7 +278,7 @@ async fn write_to_db_before_run(
 		ingress_hostnames_sorted.sort();
 
 		sql_execute!(
-			[ctx]
+			[ctx, @tx tx]
 			"
 			INSERT INTO db_job_state.run_proxied_ports (
 				run_id,
@@ -389,7 +389,7 @@ async fn bind_with_retries(
 		let port = rand::thread_rng().gen_range(range.clone()) as i32;
 
 		let (already_exists,) = sql_fetch_one!(
-			[ctx, (bool,), &mut **tx]
+			[ctx, (bool,), @tx tx]
 			"
 			SELECT EXISTS(
 				SELECT 1