+++ svc/pkg/mm/worker/src/workers/lobby_find/mod.rs
@@ -586,7 +586,7 @@ async fn insert_to_crdb(
 
 		// Insert lobby if needed
 		sql_execute!(
-			[ctx, &mut **tx]
+			[ctx, @tx tx]
 			"
 			INSERT INTO db_mm_state.lobbies (
 				lobby_id,
@@ -621,7 +621,7 @@ async fn insert_to_crdb(
 
 	// Insert query
 	sql_execute!(
-		[ctx, &mut **tx]
+		[ctx, @tx tx]
 		"
 		INSERT INTO db_mm_state.find_queries (
 			query_id,
@@ -645,7 +645,7 @@ async fn insert_to_crdb(
 	// Insert players
 	for player in players {
 		sql_execute!(
-			[ctx, &mut **tx]
+			[ctx, @tx tx]
 			"
 			INSERT INTO db_mm_state.players (
 				player_id,