+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   **Matchmaker** Expose hardware metrics to container with `RIVET_CPU`, `RIVET_MEMORY`, and `RIVET_MEMORY_OVERSUBSCRIBE`
 -   **api-cloud** `GET /cloud/bootstrapp` to provide intiial config data to the hub
 -   **api-cloud** Dynamically send Turnstile site key to hub
+-   **Infra** Rate limit on creating new SQL connections to prevent stampeding connections
 
 ### Changed